objectupdate: sched: rewrite RenderConfig
Rewrite RenderConfig to use `unstructured` vs typed API.
The rationale for this change is that doing so will
enable us to drop all the dependencies to the scheduler-plugins
API packages, which pull the scheduler API package, which
eventually pulls in k8s.io/kubernetes, which we should
really NOT depend on.

We take the change to improve the RenderConfig API and make
it forward extensible, anticipating the future support
for more scheduler plugin parameters.

The end result in terms of rendered YAML is expected
to be not 1:1 identical, but fully compatible.

Backport notice:
greatly simplified the e2e testing part (extremely simplistic)
due to different and not-backportable default changes.

Signed-off-by: Francesco Romani <fromani@redhat.com>